consistified seeking to invalid location behavior, minor fixes

+++ src/modules/sound/lullaby/VorbisDecoder.h
@@ -23,7 +23,8 @@
 
 // LOVE
 #include "common/Data.h"
-#include "Decoder.h"
+#include "common/int.h"
+#include "sound/Decoder.h"
 
 // vorbis
 #define OV_EXCLUDE_STATIC_CALLBACKS
@@ -41,8 +42,8 @@ namespace lullaby
 struct SOggFile
 {
 	const char *dataPtr;	// Pointer to the data in memory
-	int dataSize;	// Size of the data
-	int dataRead;	// How much we've read so far
+	int64 dataSize;	// Size of the data
+	int64 dataRead;	// How much we've read so far
 };
 
 class VorbisDecoder : public Decoder
